pyproject.toml
[tool.poetry]
name = "http-process-wrapper"
version = "0.1.0"
description = "A simple FastAPI project"
authors = ["Nour Selim <nourselim0@outlook.com>"]
packages = [{ include = "app" }]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = { extras = ["standard"], version = "^0.119.0" }
uvicorn = "^0.29.0"
anyio = "^4.11.0"
httpx = "^0.28.1"
PyJWT = {extras = ["crypto"], version = "^2.10.1"}
pydantic-settings = "^2.12.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
black = "^25.11.0"
isort = "^7.0.0"
pylint = "^4.0.2"

[tool.pylint]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "redefined-outer-name",
]

[tool.pylint.FORMAT]
max-line-length = 100

[tool.black]
line-length = 100

[tool.isort]
line_length = 100
profile = "black"
